Reversed direction of loop on memes.php
The loop went from 0 up, meaning that adding pictures would put them at the end of the list. To get the newest pictures at the top of the page, I reversed the loop so that it starts at the highest number and decrements to 0. Missing numbers will still break the loop as of right now.